+++ shuup/front/apps/simple_search/views.py
@@ -105,7 +105,7 @@ class SearchView(ListView):
         query = self.form.cleaned_data["q"]
         if not query:  # pragma: no cover
             return Product.objects.none()
-        return Product.objects.list_visible(self.request.shop, self.request.customer).filter(
+        return Product.objects.visible(self.request.shop, self.request.customer).filter(
             pk__in=get_search_product_ids(self.request, query))
 
     def get_context_data(self, **kwargs):